requirements.txt
flask==3.0.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-semantic-conventions==0.41b0
grpcio>=1.63.2,<2.0.0

